pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mini-pengin"
version = "0.3.0"
description = "Mini PDF pipeline (macOS-friendly): DeepSeek-OCR, Docling tables, pre/post, router."
authors = [{name="Your Name"}]
requires-python = ">=3.9"
dependencies = [
    "PyMuPDF>=1.23.4",
    "pillow>=9.5.0",
    "tqdm>=4.66.0",
    "langdetect>=1.0.9",
    "regex>=2023.10.3",
    "transformers>=4.44.0",
    "torch>=2.2.0",
    "markdown>=3.6",
    "pandas>=2.0.0",
    "lxml>=4.9.0"
]
[project.scripts]
mini-pengin = "mini_pengin.__main__:main"
